# docs/openapi.json
{
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "examples": [
          {
            "error": "Bad Request",
            "message": "Invalid input data.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message",
          "error"
        ],
        "type": "object"
      },
...
      "InternalServerError": {
        "description": "Internal Server Error",
        "examples": [
          {
            "error": "Internal Server Error.",
            "message": "An unexpected error occurred.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Internal Server Error.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "An unexpected error occurred.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the request failed due to a server error",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "NotFound": {
        "description": "Not Found",
        "examples": [
          {
            "error": "Not Found.",
            "message": "Resource not found.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Not Found.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "Resource not found.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the requested resource could not be found",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "examples": [
          {
            "error": "Unauthorized.",
            "message": "Unauthorized access.",
            "success": false
          }
        ],
        "properties": {
          "error": {
            "description": "Error details, if any",
            "example": "Unauthorized.",
            "type": "string"
          },
          "message": {
            "description": "A message providing additional information",
            "example": "Unauthorized access.",
            "type": "string"
          },
          "success": {
            "description": "Indicates the request failed due to authentication failure",
            "example": false,
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    },
    "schemas": {
...
      "IssueActivityResponse": {
        "additionalProperties": false,
        "description": "Response schema for individual issue activity details.",
        "properties": {
          "action": {
            "description": "Action performed on the issue (e.g., create, update, delete).",
            "type": "string"
          },
          "column": {
            "description": "Column or field affected by the action.",
            "type": "string"
          },
          "created_at": {
            "description": "Timestamp of the activity.",
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "description": "Email of the user who performed the action.",
            "type": "string"
          },
          "entity": {
            "description": "Entity affected by the action.",
            "type": "string"
          },
          "id": {
            "description": "Unique activity ID.",
            "type": "string"
          },
          "issue_id": {
            "description": "Associated Issue ID.",
            "type": "string"
          },
          "new_value": {
            "description": "Updated value of the field.",
            "type": "string"
          },
          "old_value": {
            "description": "Previous value of the field.",
            "type": "string"
          },
          "project_id": {
            "description": "Associated Project ID.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "issue_id",
          "project_id",
          "email",
          "action",
          "entity",
          "column",
          "old_value",
          "new_value",
          "created_at"
        ],
        "type": "object"
      },
...
      "ListIssueActivityResponse": {
        "additionalProperties": false,
        "description": "Paginated response schema for listing issue activities.",
        "properties": {
          "data": {
            "description": "List of issue activities.",
            "example": [
              {
                "action": "Updated",
                "column": "title",
                "created_at": "2025-02-07T09:45:00Z",
                "email": "user@example.com",
                "entity": "Title",
                "id": "9f1a2dbe-3bc5-45fb-97ed-8f6b33cdef12",
                "issue_id": "4d2e421e-784a-424e-ac0f-f9487e3ac95c",
                "new_value": "New Title",
                "old_value": "Old Title",
                "project_id": "a5c3d21e-4f5e-4827-b3b4-9d5c11fdfb45"
              }
            ],
            "items": {
              "$ref": "#/components/schemas/IssueActivityResponse"
            },
            "type": "array"
          },
          "meta": {
            "description": "Pagination metadata.",
            "properties": {
              "limit": {
                "description": "Number of activities per page.",
                "example": 20,
                "type": "integer"
              },
              "page": {
                "description": "Current page number.",
                "example": 1,
                "type": "integer"
              },
              "total": {
                "description": "Total number of activities.",
                "example": 50,
                "type": "integer"
              }
            },
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "List Issue Activity Response",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT Token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@projectmanagement.com",
      "name": "Project Management API Team",
      "url": "http://projectmanagement.com"
    },
    "description": "API for managing clients, projects, and issues to demonstrate the features of OpenAPI 3.1 specification.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://example.com/terms/",
    "title": "Project Management API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/project/{project_id}/issue/{issue_id}/activities": {
      "get": {
        "description": "Retrieve a list of activities for a specific issue within a project, with pagination support.",
        "operationId": "ListIssueActivitiesByID",
        "parameters": [
          {
            "description": "ID of the project to list activities from",
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ID of the issue to retrieve activities for",
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Page number for pagination",
            "in": "query",
            "name": "page",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Number of activities per page",
            "in": "query",
            "name": "page_size",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListIssueActivityResponse"
                }
              }
            },
            "description": "List of issue activities retrieved successfully"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/BadRequest"
                }
              }
            },
            "description": "Invalid pagination parameters"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Unauthorized"
                }
              }
            },
            "description": "Unauthorized to access this project or issue"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NotFound"
                }
              }
            },
            "description": "Project or issue not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List activities of an issue in a project",
        "tags": [
          "Issues"
        ]
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Test server",
      "url": "https://test.projectmanagement.com/api/v1"
    },
    {
      "description": "Development server",
      "url": "https://dev.projectmanagement.com/api/v1"
    },
    {
      "description": "Production server",
      "url": "https://api.projectmanagement.com/api/v1"
    }
  ],
...
}